Try out swapping CTRL and ALT for the thumb buttons.
I'm not quite sure if this will work out. Although I like the ergonomic positioning of the alt key, for use in my tiling window manager, I quite often need to use the CTRL key in conjunction with my first function layer. I could always use 2 hands for pressing the layer switching key as well as CTRL, but maybe with this new positioning, I will be able to press both keys with the side of my thumb. Maybe I will also revert back to the old scheme, because of the awkward positioning of the alt key which I'm using rather frequently. We'll see. I can see how someone prefers an Ergodox or Moonlander...