Create a more sane approach to the layout of the F keys.
NOTE: Before that it wasn't possible to alt+fx, needed to switch ttys in linux, because the thumb keys were occupied by the mouse buttons. This now makes the former RGB/misc. layer more populated, as well as making room for further keymaps on the layers for mousekeys and numberpad